requirements.txt
numpy==1.24.3
Pillow==10.1.0
